(style): change folder structure
