+++ anndata/experimental/__init__.py
@@ -6,6 +6,7 @@ from .pytorch import AnnLoader
 from anndata._io.specs import read_elem, write_elem, IOSpec
 from ._dispatch_io import read_dispatched, write_dispatched
 from .merge import concat_on_disk
+from .backed._io import read_backed
 
 __all__ = [
     "AnnCollection",
@@ -16,4 +17,5 @@ __all__ = [
     "write_dispatched",
     "IOSpec",
     "concat_on_disk",
+    "read_backed",
 ]
